noteman/requirements.txt
psycopg
Django
djangorestframework
django-autocomplete-light
drf-spectacular
django-filter
whitenoise
